Move the translation context defs to translate_ctx.rs
