+++ charon/src/charon-driver.rs
@@ -68,6 +68,7 @@ mod remove_unused_locals;
 mod reorder_decls;
 mod rust_to_local_ids;
 mod translate_constants;
+mod translate_ctx;
 mod translate_functions_to_ullbc;
 mod translate_types;
 mod types;
