+++ charon/src/lib.rs
@@ -78,6 +78,7 @@ pub mod remove_unused_locals;
 pub mod reorder_decls;
 pub mod rust_to_local_ids;
 pub mod translate_constants;
+pub mod translate_ctx;
 pub mod translate_functions_to_ullbc;
 pub mod translate_types;
 pub mod types;
