+++ charon/src/translate_functions_to_ullbc.rs
@@ -22,9 +22,10 @@ use crate::translate_constants::{
     translate_const_kind_as_const_generic, translate_evaluated_operand_constant,
     translate_operand_constant,
 };
+use crate::translate_ctx::*;
 use crate::translate_types;
 use crate::types as ty;
-use crate::types::{FieldId, LiteralTy, VariantId};
+use crate::types::{FieldId, VariantId};
 use crate::ullbc_ast as ast;
 use crate::values as v;
 use crate::values::{Literal, ScalarValue};
@@ -43,297 +44,7 @@ use rustc_session::Session;
 use rustc_span::Span;
 use std::collections::HashMap;
 use std::iter::FromIterator;
-use translate_types::{translate_erased_region, translate_region_name, TypeTransContext};
-
-/// Translation context for function and global definitions
-/// TODO: we should put the `TyCtx` and the `Session` in there
-pub struct DeclTransContext<'tcx, 'ctx> {
-    /// The compiler session
-    pub sess: &'ctx Session,
-    ///
-    pub tcx: TyCtxt<'tcx>,
-    /// The ordered declarations
-    pub ordered: &'ctx OrderedDecls,
-    /// Type definitions
-    pub type_defs: &'ctx ty::TypeDecls,
-    /// The function definitions
-    pub fun_defs: &'ctx ast::FunDecls,
-    /// The global definitions
-    pub global_defs: &'ctx ast::GlobalDecls,
-    /// The level at which to extract the MIR
-    pub mir_level: MirLevel,
-}
-
-/// A translation context for function and global bodies.
-///
-/// We use `im::OrdMap` a lot in places where we could use a
-/// `std::collections::BTreeMap` (because we actually don't do context
-/// duplication). For now the performance doesn't matter, and it allows
-/// not to mix too many very similar data structures, but we might consider
-/// using `std::collections::BTreeMap` instead.
-#[derive(Clone)]
-struct BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    /// This is used in very specific situations.
-    def_id: DefId,
-    /// The declarations translation context, containing the function and global definitions.
-    /// Also contains the type translation context.
-    ft_ctx: &'ctx DeclTransContext<'tcx, 'ctx1>,
-    /// Region counter
-    regions_counter: ty::RegionVarId::Generator,
-    /// The regions
-    regions: ty::RegionVarId::Vector<ty::RegionVar>,
-    /// The map from rust region to translated region indices
-    rregions_to_ids: im::OrdMap<rustc_middle::ty::RegionKind<'tcx>, ty::RegionVarId::Id>,
-    /// Id counter for the type variables
-    type_vars_counter: ty::TypeVarId::Generator,
-    /// The type variables
-    type_vars: ty::TypeVarId::Vector<ty::TypeVar>,
-    /// The map from rust type variable indices to translated type variable
-    /// indices.
-    rtype_vars_to_ids: im::OrdMap<u32, ty::TypeVarId::Id>,
-    /// Id counter for the variables
-    vars_counter: v::VarId::Generator,
-    /// The "regular" variables
-    vars: v::VarId::Vector<ast::Var>,
-    /// The map from rust variable indices to translated variables indices.
-    rvars_to_ids: im::OrdMap<u32, v::VarId::Id>,
-    /// Id counter for the const generic variables
-    const_generic_counter: ty::ConstGenericVarId::Generator,
-    /// The const generic variables
-    const_generic_vars: ty::ConstGenericVarId::Vector<ty::ConstGenericVar>,
-    /// The map from rust const generic variables to translate const generic
-    /// variable indices.
-    rconst_generic_vars_to_ids: im::OrdMap<u32, ty::ConstGenericVarId::Id>,
-    /// Block id counter
-    blocks_counter: ast::BlockId::Generator,
-    /// The translated blocks. We can't use `ast::BlockId::Vector<ast::BlockData>`
-    /// here because we might generate several fresh indices before actually
-    /// adding the resulting blocks to the map.
-    blocks: im::OrdMap<ast::BlockId::Id, ast::BlockData>,
-    /// The map from rust blocks to translated blocks.
-    /// Note that when translating terminators like DropAndReplace, we might have
-    /// to introduce new blocks which don't appear in the original MIR.
-    rblocks_to_ids: im::OrdMap<BasicBlock, ast::BlockId::Id>,
-}
-
-impl<'tcx, 'ctx> DeclTransContext<'tcx, 'ctx> {
-    fn get_def_id_from_rid(&self, def_id: DefId) -> Option<ast::FunDeclId::Id> {
-        self.ordered.fun_rid_to_id.get(&def_id).copied()
-    }
-
-    fn get_def_rid_from_id(&self, def_id: ast::FunDeclId::Id) -> Option<DefId> {
-        self.ordered
-            .decls_info
-            .get(&AnyDeclId::Fun(def_id))
-            .map(|i| i.rid)
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    /// Create a new `ExecContext`.
-    fn new(def_id: DefId, ft_ctx: &'ctx DeclTransContext<'tcx, 'ctx1>) -> Self {
-        BodyTransContext {
-            def_id,
-            ft_ctx,
-            regions_counter: ty::RegionVarId::Generator::new(),
-            regions: ty::RegionVarId::Vector::new(),
-            rregions_to_ids: im::OrdMap::new(),
-            type_vars_counter: ty::TypeVarId::Generator::new(),
-            type_vars: ty::TypeVarId::Vector::new(),
-            rtype_vars_to_ids: im::OrdMap::new(),
-            vars_counter: v::VarId::Generator::new(),
-            vars: v::VarId::Vector::new(),
-            rvars_to_ids: im::OrdMap::new(),
-            const_generic_counter: ty::ConstGenericVarId::Generator::new(),
-            const_generic_vars: ty::ConstGenericVarId::Vector::new(),
-            rconst_generic_vars_to_ids: im::OrdMap::new(),
-            blocks_counter: ast::BlockId::Generator::new(),
-            blocks: im::OrdMap::new(),
-            rblocks_to_ids: im::OrdMap::new(),
-        }
-    }
-
-    fn get_local(&self, local: &mir::Local) -> Option<v::VarId::Id> {
-        self.rvars_to_ids.get(&local.as_u32()).copied()
-    }
-
-    fn get_block_id_from_rid(&self, rid: BasicBlock) -> Option<ast::BlockId::Id> {
-        self.rblocks_to_ids.get(&rid).copied()
-    }
-
-    fn get_var_from_id(&self, var_id: v::VarId::Id) -> Option<&ast::Var> {
-        self.vars.get(var_id)
-    }
-
-    fn get_region_from_rust(
-        &self,
-        r: rustc_middle::ty::RegionKind<'tcx>,
-    ) -> Option<ty::RegionVarId::Id> {
-        self.rregions_to_ids.get(&r).copied()
-    }
-
-    fn push_region(
-        &mut self,
-        r: rustc_middle::ty::RegionKind<'tcx>,
-        name: Option<String>,
-    ) -> ty::RegionVarId::Id {
-        use crate::id_vector::ToUsize;
-        let rid = self.regions_counter.fresh_id();
-        assert!(rid.to_usize() == self.regions.len());
-        let var = ty::RegionVar { index: rid, name };
-        self.regions.insert(rid, var);
-        self.rregions_to_ids.insert(r, rid);
-        rid
-    }
-
-    fn push_type_var(&mut self, rindex: u32, name: String) -> ty::TypeVarId::Id {
-        use crate::id_vector::ToUsize;
-        let var_id = self.type_vars_counter.fresh_id();
-        assert!(var_id.to_usize() == self.type_vars.len());
-        let var = ty::TypeVar {
-            index: var_id,
-            name,
-        };
-        self.type_vars.insert(var_id, var);
-        self.rtype_vars_to_ids.insert(rindex, var_id);
-        var_id
-    }
-
-    fn push_var(&mut self, rid: u32, ty: ty::ETy, name: Option<String>) {
-        use crate::id_vector::ToUsize;
-        let var_id = self.vars_counter.fresh_id();
-        assert!(var_id.to_usize() == self.vars.len());
-        let var = ast::Var {
-            index: var_id,
-            name,
-            ty,
-        };
-        self.vars.insert(var_id, var);
-        self.rvars_to_ids.insert(rid, var_id);
-    }
-
-    fn push_const_generic_var(&mut self, rid: u32, ty: LiteralTy, name: String) {
-        use crate::id_vector::ToUsize;
-        let var_id = self.const_generic_counter.fresh_id();
-        assert!(var_id.to_usize() == self.vars.len());
-        let var = ty::ConstGenericVar {
-            index: var_id,
-            name,
-            ty,
-        };
-        self.const_generic_vars.insert(var_id, var);
-        self.rconst_generic_vars_to_ids.insert(rid, var_id);
-    }
-
-    fn fresh_block_id(&mut self, rid: BasicBlock) -> ast::BlockId::Id {
-        let block_id = self.blocks_counter.fresh_id();
-        self.rblocks_to_ids.insert(rid, block_id);
-        block_id
-    }
-
-    fn push_block(&mut self, id: ast::BlockId::Id, block: ast::BlockData) {
-        self.blocks.insert(id, block);
-    }
-
-    fn get_type_defs(&self) -> &ty::TypeDecls {
-        self.ft_ctx.type_defs
-    }
-}
-
-impl<'tcx, 'ctx> Formatter<ty::TypeDeclId::Id> for DeclTransContext<'tcx, 'ctx> {
-    fn format_object(&self, id: ty::TypeDeclId::Id) -> String {
-        self.type_defs.format_object(id)
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<ty::TypeVarId::Id> for BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    fn format_object(&self, id: ty::TypeVarId::Id) -> String {
-        let v = self.type_vars.get(id).unwrap();
-        v.to_string()
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<v::VarId::Id> for BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    fn format_object(&self, id: v::VarId::Id) -> String {
-        let v = self.vars.get(id).unwrap();
-        v.to_string()
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<ty::RegionVarId::Id> for BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    fn format_object(&self, id: ty::RegionVarId::Id) -> String {
-        let v = self.regions.get(id).unwrap();
-        v.to_string()
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<&ty::Region<ty::RegionVarId::Id>>
-    for BodyTransContext<'tcx, 'ctx, 'ctx1>
-{
-    fn format_object(&self, r: &ty::Region<ty::RegionVarId::Id>) -> String {
-        r.fmt_with_ctx(self)
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<ty::ConstGenericVarId::Id>
-    for BodyTransContext<'tcx, 'ctx, 'ctx1>
-{
-    fn format_object(&self, id: ty::ConstGenericVarId::Id) -> String {
-        let v = self.const_generic_vars.get(id).unwrap();
-        v.to_string()
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<&ty::ErasedRegion> for BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    fn format_object(&self, _: &ty::ErasedRegion) -> String {
-        "'_".to_owned()
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<ty::TypeDeclId::Id> for BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    fn format_object(&self, id: ty::TypeDeclId::Id) -> String {
-        self.ft_ctx.type_defs.format_object(id)
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<ty::GlobalDeclId::Id> for BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    fn format_object(&self, id: ty::GlobalDeclId::Id) -> String {
-        self.ft_ctx.global_defs.format_object(id)
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<&ty::Ty<ty::Region<ty::RegionVarId::Id>>>
-    for BodyTransContext<'tcx, 'ctx, 'ctx1>
-{
-    fn format_object(&self, ty: &ty::Ty<ty::Region<ty::RegionVarId::Id>>) -> String {
-        ty.fmt_with_ctx(self)
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> Formatter<&ty::Ty<ty::ErasedRegion>>
-    for BodyTransContext<'tcx, 'ctx, 'ctx1>
-{
-    fn format_object(&self, ty: &ty::Ty<ty::ErasedRegion>) -> String {
-        ty.fmt_with_ctx(self)
-    }
-}
-
-impl<'tcx, 'ctx, 'ctx1> BodyTransContext<'tcx, 'ctx, 'ctx1> {
-    fn to_type_trans_context(&self) -> TypeTransContext<'tcx, 'ctx> {
-        TypeTransContext::new(
-            // For the def id: we use the function/global def id
-            self.def_id,
-            self.ft_ctx.tcx,
-            self.ft_ctx.type_defs,
-            self.ft_ctx.global_defs,
-            self.ft_ctx.ordered,
-            self.rregions_to_ids.clone(),
-            self.rtype_vars_to_ids.clone(),
-            self.rconst_generic_vars_to_ids.clone(),
-            self.ft_ctx.mir_level,
-        )
-    }
-}
+use translate_types::{translate_erased_region, translate_region_name};
 
 fn translate_ety<'tcx>(
     bt_ctx: &BodyTransContext<'tcx, '_, '_>,
