+++ charon/src/translate_types.rs
@@ -11,164 +11,15 @@ use crate::regions_hierarchy::TypesConstraintsMap;
 use crate::reorder_decls::DeclarationGroup;
 use crate::rust_to_local_ids::*;
 use crate::translate_constants::translate_const_kind_as_const_generic;
+use crate::translate_ctx::*;
 use crate::types as ty;
-use crate::types::{ConstGeneric, TypeDeclId};
+use crate::types::ConstGeneric;
 use crate::ullbc_ast as ast;
 use core::convert::*;
 use rustc_hir::def_id::DefId;
 use rustc_middle::mir::Mutability;
 use rustc_middle::ty::{Ty, TyCtxt, TyKind};
 use rustc_session::Session;
-
-/// Translation context for type definitions
-/// TODO: merge with BodyTransContext
-#[derive(Clone)]
-pub struct TypeTransContext<'tcx, 'ctx> {
-    /// This is used in very specific situations, for instance to retrieve parameter
-    /// environments in [crate::translate_constants]. This should be the id of the
-    /// definition we are exploring (can be the id of a type or function definition, for
-    /// instance).
-    pub def_id: DefId,
-    ///
-    pub tcx: TyCtxt<'tcx>,
-    /// The type definitions
-    pub type_defs: &'ctx ty::TypeDecls,
-    /// The global definitions
-    pub global_defs: &'ctx ast::GlobalDecls,
-    /// Ordered declarations allowing to convert id to and from rid.
-    pub ordered: &'ctx OrderedDecls,
-    /// The map from Rust region var to LLBC region var
-    pub region_params_map: im::OrdMap<rustc_middle::ty::RegionKind<'tcx>, ty::RegionVarId::Id>,
-    /// The map from Rust type var identifier to LLBC type var
-    pub type_params_map: im::OrdMap<u32, ty::TypeVarId::Id>,
-    /// The map from Rust const generic var identifiers to LLBC const generic vars
-    pub const_generic_params_map: im::OrdMap<u32, ty::ConstGenericVarId::Id>,
-    /// The level at which to extract the MIR
-    pub mir_level: MirLevel,
-}
-
-impl<'tcx, 'ctx> TypeTransContext<'tcx, 'ctx> {
-    pub fn new(
-        def_id: DefId,
-        tcx: TyCtxt<'tcx>,
-        type_defs: &'ctx ty::TypeDecls,
-        global_defs: &'ctx ast::GlobalDecls,
-        ordered: &'ctx OrderedDecls,
-        region_params_map: im::OrdMap<rustc_middle::ty::RegionKind<'tcx>, ty::RegionVarId::Id>,
-        type_params_map: im::OrdMap<u32, ty::TypeVarId::Id>,
-        const_generic_params_map: im::OrdMap<u32, ty::ConstGenericVarId::Id>,
-        mir_level: MirLevel,
-    ) -> Self {
-        Self {
-            def_id,
-            tcx,
-            type_defs,
-            global_defs,
-            ordered,
-            region_params_map,
-            type_params_map,
-            const_generic_params_map,
-            mir_level,
-        }
-    }
-
-    pub fn get_id(&self, rid: DefId) -> TypeDeclId::Id {
-        *self.ordered.type_rid_to_id.get(&rid).unwrap()
-    }
-}
-
-/// Auxiliary definition used to format definitions.
-struct TypeDeclFormatter<'a> {
-    type_defs: &'a ty::TypeDecls,
-    global_defs: &'a ast::GlobalDecls,
-    /// The region parameters of the definition we are printing (needed to
-    /// correctly pretty print region var ids)
-    region_params: &'a ty::RegionVarId::Vector<ty::RegionVar>,
-    /// The type parameters of the definition we are printing (needed to
-    /// correctly pretty print type var ids)
-    type_params: &'a ty::TypeVarId::Vector<ty::TypeVar>,
-    /// The const generic parameters of the definition we are printing (needed to
-    /// correctly pretty print type var ids)
-    const_generic_params: &'a ty::ConstGenericVarId::Vector<ty::ConstGenericVar>,
-}
-
-impl<'tcx, 'ctx> Formatter<ty::TypeDeclId::Id> for TypeTransContext<'tcx, 'ctx> {
-    fn format_object(&self, id: ty::TypeDeclId::Id) -> String {
-        self.type_defs.format_object(id)
-    }
-}
-
-impl<'a> Formatter<ty::RegionVarId::Id> for TypeDeclFormatter<'a> {
-    fn format_object(&self, id: ty::RegionVarId::Id) -> String {
-        // Lookup the region parameter
-        let v = self.region_params.get(id).unwrap();
-        // Format
-        v.to_string()
-    }
-}
-
-impl<'a> Formatter<ty::ConstGenericVarId::Id> for TypeDeclFormatter<'a> {
-    fn format_object(&self, id: ty::ConstGenericVarId::Id) -> String {
-        // Lookup the region parameter
-        let v = self.const_generic_params.get(id).unwrap();
-        // Format
-        v.to_string()
-    }
-}
-
-impl<'a> Formatter<ty::TypeVarId::Id> for TypeDeclFormatter<'a> {
-    fn format_object(&self, id: ty::TypeVarId::Id) -> String {
-        // Lookup the type parameter
-        let v = self.type_params.get(id).unwrap();
-        // Format
-        v.to_string()
-    }
-}
-
-impl<'a> Formatter<&ty::Region<ty::RegionVarId::Id>> for TypeDeclFormatter<'a> {
-    fn format_object(&self, r: &ty::Region<ty::RegionVarId::Id>) -> String {
-        r.fmt_with_ctx(self)
-    }
-}
-
-impl<'a> Formatter<&ty::ErasedRegion> for TypeDeclFormatter<'a> {
-    fn format_object(&self, _: &ty::ErasedRegion) -> String {
-        "".to_owned()
-    }
-}
-
-impl<'a> Formatter<&ty::TypeDecl> for TypeDeclFormatter<'a> {
-    fn format_object(&self, def: &ty::TypeDecl) -> String {
-        def.fmt_with_ctx(self)
-    }
-}
-
-impl<'a> Formatter<ty::TypeDeclId::Id> for TypeDeclFormatter<'a> {
-    fn format_object(&self, id: ty::TypeDeclId::Id) -> String {
-        self.type_defs.format_object(id)
-    }
-}
-
-impl<'a> Formatter<ty::GlobalDeclId::Id> for TypeDeclFormatter<'a> {
-    fn format_object(&self, id: ty::GlobalDeclId::Id) -> String {
-        self.global_defs.format_object(id)
-    }
-}
-
-impl<'tcx, 'ctx> Formatter<&ty::TypeDecl> for TypeTransContext<'tcx, 'ctx> {
-    fn format_object(&self, def: &ty::TypeDecl) -> String {
-        // Create a type def formatter (which will take care of the
-        // type parameters)
-        let formatter = TypeDeclFormatter {
-            type_defs: self.type_defs,
-            global_defs: self.global_defs,
-            region_params: &def.region_params,
-            type_params: &def.type_params,
-            const_generic_params: &def.const_generic_params,
-        };
-        formatter.format_object(def)
-    }
-}
 
 pub fn translate_region_name(region: &rustc_middle::ty::RegionKind<'_>) -> Option<String> {
     // Compute the region name
